qwen-chat/requirements.txt
llama-cpp-python>=0.3.0
huggingface-hub>=0.25.0
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.0.0
rich>=13.0.0
